refactor: Alert 메서드를 UIViewController 메서드로 분리
